requirements.txt
flask==2.0.1
werkzeug==2.0.3
# spacy==3.5.0
sqlalchemy==1.4.23
python-dotenv==0.19.0
flask-sqlalchemy==2.5.1
google-generativeai==0.3.1
pandas==2.0.3
openpyxl==3.1.2